Start moving API routes into appropriate modules
