+++ server/main.ts
@@ -4,11 +4,11 @@ import { cors } from "./cors.js";
 import { db } from "./db.js";
 import { env } from "./environment.js";
 import { handleErrors } from "./handleErrors.js";
-import { lol } from "./lol.js";
+import { lol } from "./routes/v0/index.js";
 import { NotFoundError } from "./errors/NotFoundError.js";
-import { ping } from "./ping.js";
+import { ping } from "./routes/v0/ping.js";
 import { respondError } from "./responses.js";
-import { serverVersion } from "./serverVersion.js";
+import { serverVersion } from "./routes/v0/version.js";
 import { session } from "./auth/jwt.js";
 import { version as appVersion } from "./version.js";
 // import csurf from "csurf"; // TODO: Might be important later
@@ -16,25 +16,24 @@ import express from "express";
 import expressWs from "express-ws";
 import helmet from "helmet";
 
-const API_VERSION = 0;
 const PORT = 40850;
 
-const app = express()
+export const app = express()
 	.use(helmet()) // also disables 'x-powered-by' header
 	.use(cors());
 
 expressWs(app); // Set up websocket support. This is the reason our endpoint declarations need to be functions and not `const` declarations
 
 app
-	.get(`/v${API_VERSION}/`, lol)
-	.get(`/v${API_VERSION}/ping`, ping)
-	.get(`/v${API_VERSION}/version`, serverVersion)
+	.get("/v0/", lol)
+	.get("/v0/ping", ping)
+	.get("/v0/version", serverVersion)
 	.set("trust proxy", 1) // trust first proxy
 	.use(session)
 	.use(express.json({ limit: "5mb" }))
 	.use(express.urlencoded({ limit: "5mb", extended: true }))
-	.use(`/v${API_VERSION}/`, auth()) // Auth endpoints
-	.use(`/v${API_VERSION}/db`, db()) // Database endpoints (checks auth)
+	.use("/v0/", auth()) // Auth endpoints
+	.use("/v0/db", db()) // Database endpoints (checks auth)
 	.use((req, res) => {
 		// Custom 404
 		respondError(res, new NotFoundError());
