+++ server/routes/v0/ping.ts
