+++ server/routes/v0/version.ts
@@ -1,5 +1,5 @@
 import type { RequestHandler } from "express";
-import { version } from "./version.js";
+import { version } from "../../version.js";
 
 export const serverVersion: RequestHandler = (req, res) =>
 	res.json({ message: `Accountable v${version}`, version });
